cambio variables float por double

+++ tarea1/src/ar/edu/unlam/tallerweb/tarea1/Circulo.java
@@ -1,15 +1,14 @@
 package ar.edu.unlam.tallerweb.tarea1;
 
 public class Circulo {
-	public Float radio = (float)6;
-	public Float pi = (float)Math.PI;
-	
-	public float perimetro(){
+	public Double radio = 6.00;
+	public Double pi = Math.PI;
+
+	public Double perimetro() {
 		return (2 * pi * radio);
 	}
-	
-	public float area(){
-		return( pi * (radio*radio));
+
+	public Double area() {
+		return (pi * (radio * radio));
 	}
-	
 }
